refactor: rename Android Java package to com.nitromelondb
Drop the leftover com.nozbe.watermelondb / com.nozbe.watermelonTest
namespaces so the library and test harness match the nitromelondb identity.
JNI entry points, FindClass, Gradle namespace, Proguard keep rules, and
install/migrate docs follow the same rename.
