+++ docs-website/docs/docs/Installation.mdx
@@ -157,10 +157,10 @@ If you previously installed JSI manually, remove:
 
 </details>
 
-If you minify with R8 / Proguard, keep this rule. The Android Java package is still `com.nozbe.watermelondb`:
+If you minify with R8 / Proguard, keep this rule:
 
 ```
--keep class com.nozbe.watermelondb.** { *; }
+-keep class com.nitromelondb.** { *; }
 ```
 
 ## Web setup
