+++ docs-website/docs/docs/Migrating.md
@@ -137,13 +137,13 @@ That path **does not exist**. Remove the whole JSI block. Do not point it at `na
 
 Do not register `WatermelonDBPackage` by hand — that is the old architecture. Autolinking is required.
 
-The Android Java package is still `com.nozbe.watermelondb`. If you already have this R8 / Proguard rule, **leave it** — do not delete it:
+The Android Java package is `com.nitromelondb` (it used to be `com.nozbe.watermelondb`). Update any R8 / Proguard keep rule:
 
 ```
--keep class com.nozbe.watermelondb.** { *; }
+-keep class com.nitromelondb.** { *; }
 ```
 
-Turbo-sync JSON injection now goes through `com.nozbe.watermelondb.NitromelonNative.provideSyncJson`.
+Turbo-sync JSON injection now goes through `com.nitromelondb.NitromelonNative.provideSyncJson`.
 
 Minimum Android SDK is **24**.
