+++ native/android/src/main/java/com/nitromelondb/WatermelonDBPackage.java
@@ -1,4 +1,4 @@
-package com.nozbe.watermelondb;
+package com.nitromelondb;
 
 import androidx.annotation.NonNull;
