+++ native/androidTest/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.nozbe.watermelondb",
+    package = "com.nitromelondb",
 )
 
 android_resource(
     name = "res",
-    package = "com.nozbe.watermelondb",
+    package = "com.nitromelondb",
     res = "src/main/res",
 )
