+++ native/androidTest/app/src/main/java/com/nitromelondb/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.nozbe.watermelonTest
+package com.nitromelondb
 
 import android.app.Application
 import com.facebook.react.PackageList
@@ -6,7 +6,6 @@ import com.facebook.react.ReactApplication
 import com.facebook.react.ReactHost
 import com.facebook.react.ReactNativeApplicationEntryPoint.loadReactNative
 import com.facebook.react.defaults.DefaultReactHost.getDefaultReactHost
-import com.nozbe.watermelondb.WatermelonDBPackage
 
 class MainApplication : Application(), ReactApplication {
