+++ native/androidTest/app/src/main/java/com/nitromelondb/NativeModulesPackage.kt
@@ -1,4 +1,4 @@
-package com.nozbe.watermelonTest
+package com.nitromelondb
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
